v1.99h (stable) 17.02.2018 - Imported UDP local network broadcast message implementation for uIP.c (from v3.0 source tree) - Imported ksz8851 network chip data offset patch for NetSDR.c (from v3.0 source tree) - All USB endpoints are now being refreshed even if DATAMODE_NETWORK (from v3.0 source tree) - Included LUFA12019 to source tree
This is the last v1.xx branch. The changes here are actually retrofits from the 1.99g build
to remedy the network stalling problem. However, as we managed to make a bug in 2.0 code
tree what took 5 (!!!) years to discover, some of the most important fixes were propagated
backwards on 2018, so we would have at least some improvement.


git-svn-id: https://localhost/svn/sdr-mk15/trunk@17 d1d4a745-7860-4ac4-97b5-893427978f15